authRotes & authController, fixes for cookie error
