+++ server/src/index.ts
@@ -1,5 +1,6 @@
 import express, { Application } from "express";
 import Server from "./server";
+import authRouter from "./routes/auth.routes";
 import userRouter from "./routes/user.routes";
 import rolesRouter from "./routes/roles.routes";
 import privelegesRouter from "./routes/priveleges.routes";
@@ -12,12 +13,13 @@ const app: Application = express();
 const server: Server = new Server(app);
 const PORT: number = process.env.PORT ? parseInt(process.env.PORT, 10) : 8080;
 
-app.use("/api/users", userRouter);
+app.use("/api", authRouter);
 app.use("/api/catalog", catalogRouter);
 app.use("/api/product-types", productTypesRouter);
 
 app.use(verifyToken);
 app.use(verifyRole);
+app.use("/api/users", userRouter);
 app.use("/api/roles", rolesRouter);
 app.use("/api/priveleges", privelegesRouter);
